requirements.txt
langchain
langchain-community
crewai
crewai-tools
openai
pandas
exa-py
python-dotenv
setuptools
yfinance
unstructured
requests_html
sec_api
faiss-cpu
quickfs
matplotlib